requirements.txt
requests
kafka-python
bs4
numpy
pandas
simplejson
websockets